Refactor: Rename Habit to UsefulHabit and move mappers
- Renamed `Habit` model and related classes/files to `UsefulHabit` for clarity.
- Moved Note mapper functions to a dedicated `mappers` package.
- Updated `UsefulHabitsScreen` to display a sample `UsefulHabitCard`.
- Introduced `UsefulHabitDto`.
